refactor(workspace): extract window tab state helpers
